Move STM32-specific ADC pinouts to ports/stm32/stm32_common.cpp
